+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 > *(under development)*
 <!------------------------------------------------------------------------------------------------->
 
+- **new**:
+    - extend `OptimizationStrategy` implementation to support scheduled parameters, that evolve according to a schedule during multi-iteration execution of a strategy 
+
 - **improved**:
     - tweak layout of solver benchmark docs